+++ myproject/config.py
@@ -3,7 +3,7 @@ import pymysql
 DB_CONFIG = {
     'host': 'localhost',
     'user': 'root',
-    'password': 'comep1522w',
-    'database': 'sm',
+    'password': 'q1w2e3r4^',
+    'database': 'flask_db',
     'charset': 'utf8mb4'
 }
