+++ myproject/docs/db.txt
